+++ .github/workflows/ci.yml
@@ -564,9 +564,9 @@ jobs:
           python_version=${{ matrix.python-version }}
           wheelfile=$(ls ./wheels/scikit_decide*-cp${python_version/\./}-*win*.whl)
           if [ "$python_version" = "3.12" ]; then
-            pip install ${wheelfile}[all] pytest "pygame>=2.5" optuna "cffi>=1.17" graph-jsp-env pytest-cases "matplotlib<3.10"
+            pip install ${wheelfile}[all] pytest "pygame>=2.5" optuna "cffi>=1.17" pytest-cases "matplotlib<3.10" "git+https://github.com/nhuet/graph-jsp-env.git@signal-rllib"
           else
-            pip install ${wheelfile}[all] pytest gymnasium[classic-control] optuna graph-jsp-env pytest-cases
+            pip install ${wheelfile}[all] pytest gymnasium[classic-control] optuna pytest-cases "git+https://github.com/nhuet/graph-jsp-env.git@signal-rllib"
           fi
 
       - name: Test with pytest
@@ -664,9 +664,9 @@ jobs:
           arch=$(uname -m)
           wheelfile=$(ls ./wheels/scikit_decide*-cp${python_version/\./}-*macos*${arch}.whl)
           if [ "$python_version" = "3.12" ]; then
-            pip install ${wheelfile}[all] pytest "pygame>=2.5" optuna "cffi>=1.17" graph-jsp-env pytest-cases "matplotlib<3.10"
+            pip install ${wheelfile}[all] pytest "pygame>=2.5" optuna "cffi>=1.17" pytest-cases "matplotlib<3.10" "git+https://github.com/nhuet/graph-jsp-env.git@signal-rllib"
           else
-            pip install ${wheelfile}[all] pytest gymnasium[classic-control] optuna graph-jsp-env pytest-cases
+            pip install ${wheelfile}[all] pytest gymnasium[classic-control] optuna pytest-cases "git+https://github.com/nhuet/graph-jsp-env.git@signal-rllib"
           fi
 
       - name: Test with pytest
@@ -764,9 +764,9 @@ jobs:
           python_version=${{ matrix.python-version }}
           wheelfile=$(ls ./wheels/scikit_decide*-cp${python_version/\./}-*manylinux*.whl)
           if [ "$python_version" = "3.12" ]; then
-            pip install ${wheelfile}[all] pytest "pygame>=2.5" "cffi>=1.17" docopt commonmark optuna graph-jsp-env pytest-cases "matplotlib<3.10"
+            pip install ${wheelfile}[all] pytest "pygame>=2.5" "cffi>=1.17" docopt commonmark optuna pytest-cases "matplotlib<3.10" "git+https://github.com/nhuet/graph-jsp-env.git@signal-rllib"
           else
-            pip install ${wheelfile}[all] pytest gymnasium[classic-control] docopt commonmark optuna graph-jsp-env pytest-cases
+            pip install ${wheelfile}[all] pytest gymnasium[classic-control] docopt commonmark optuna pytest-cases "git+https://github.com/nhuet/graph-jsp-env.git@signal-rllib"
           fi
 
       - name: Test with pytest
